+++ src/dvg_devices/Keysight_N8700_demo.py
@@ -19,7 +19,7 @@ from dvg_utils.dvg_pyqt_controls import SS_TEXTBOX_READ_ONLY
 
 from dvg_devices.Keysight_N8700_protocol_SCPI import Keysight_N8700
 from dvg_devices.Keysight_N8700_qdev import Keysight_N8700_qdev
-from dvg_qdeviceio import DAQ_trigger
+from dvg_qdeviceio import DAQ_TRIGGER
 
 # Show debug info in terminal? Warning: Slow! Do not leave on unintentionally.
 DEBUG = False
@@ -27,6 +27,7 @@ DEBUG = False
 # ------------------------------------------------------------------------------
 #   MainWindow
 # ------------------------------------------------------------------------------
+
 
 class MainWindow(QtWid.QWidget):
     def __init__(self, parent=None, **kwargs):
@@ -36,8 +37,10 @@ class MainWindow(QtWid.QWidget):
         self.setWindowTitle("Keysight N8700 power supply control")
 
         # Top grid
-        self.lbl_title = QtWid.QLabel("Keysight N8700 power supply control",
-            font=QtGui.QFont("Palatino", 14, weight=QtGui.QFont.Bold))
+        self.lbl_title = QtWid.QLabel(
+            "Keysight N8700 power supply control",
+            font=QtGui.QFont("Palatino", 14, weight=QtGui.QFont.Bold),
+        )
         self.pbtn_exit = QtWid.QPushButton("Exit")
         self.pbtn_exit.clicked.connect(self.close)
         self.pbtn_exit.setMinimumHeight(30)
@@ -58,9 +61,11 @@ class MainWindow(QtWid.QWidget):
         vbox.addLayout(hbox1)
         vbox.addStretch(1)
 
+
 # ------------------------------------------------------------------------------
 #   about_to_quit
 # ------------------------------------------------------------------------------
+
 
 def about_to_quit():
     print("About to quit")
@@ -68,19 +73,27 @@ def about_to_quit():
     for psu_qdev in psus_qdev:
         psu_qdev.quit()
     for psu in psus:
-        try: psu.close()
-        except: pass
-    try: rm.close()
-    except: pass
+        try:
+            psu.close()
+        except:
+            pass
+    try:
+        rm.close()
+    except:
+        pass
+
 
 # ------------------------------------------------------------------------------
 #   trigger_update_psus
 # ------------------------------------------------------------------------------
 
+
 def trigger_update_psus():
-    if DEBUG: dprint("timer_psus: wake up all DAQ")
+    if DEBUG:
+        dprint("timer_psus: wake up all DAQ")
     for psu_qdev in psus_qdev:
         psu_qdev.worker_DAQ.wake_up()
+
 
 # ------------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
@@ -90,22 +103,25 @@ def trigger_update_psus():
 # ------------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     # VISA addresses of the Keysight PSUs
     VISA_ADDRESS_PSU_1 = "USB0::0x0957::0x8707::US15M3727P::INSTR"
     VISA_ADDRESS_PSU_2 = "USB0::0x0957::0x8707::US15M3728P::INSTR"
     VISA_ADDRESS_PSU_3 = "USB0::0x0957::0x8707::US15M3726P::INSTR"
 
     # Config files
-    PATH_CONFIG_PSU_1 = Path(os.getcwd() +
-                             "/config/settings_Keysight_PSU_1.txt")
-    PATH_CONFIG_PSU_2 = Path(os.getcwd() +
-                             "/config/settings_Keysight_PSU_2.txt")
-    PATH_CONFIG_PSU_3 = Path(os.getcwd() +
-                             "/config/settings_Keysight_PSU_3.txt")
+    PATH_CONFIG_PSU_1 = Path(
+        os.getcwd() + "/config/settings_Keysight_PSU_1.txt"
+    )
+    PATH_CONFIG_PSU_2 = Path(
+        os.getcwd() + "/config/settings_Keysight_PSU_2.txt"
+    )
+    PATH_CONFIG_PSU_3 = Path(
+        os.getcwd() + "/config/settings_Keysight_PSU_3.txt"
+    )
 
     # The state of the PSUs is polled with this time interval
-    UPDATE_INTERVAL_MS = 1000       # [ms]
+    UPDATE_INTERVAL_MS = 1000  # [ms]
 
     # --------------------------------------------------------------------------
     #   Connect to and set up Keysight power supplies (PSU)
@@ -126,9 +142,9 @@ if __name__ == '__main__':
     # --------------------------------------------------------------------------
     #   Create application
     # --------------------------------------------------------------------------
-    QtCore.QThread.currentThread().setObjectName('MAIN')    # For DEBUG info
+    QtCore.QThread.currentThread().setObjectName("MAIN")  # For DEBUG info
 
-    app = 0    # Work-around for kernel crash when using Spyder IDE
+    app = 0  # Work-around for kernel crash when using Spyder IDE
     app = QtWid.QApplication(sys.argv)
     app.setFont(QtGui.QFont("Arial", 9))
     app.setStyleSheet(SS_TEXTBOX_READ_ONLY)
@@ -140,15 +156,18 @@ if __name__ == '__main__':
 
     psus_qdev = list()
     for i in range(len(psus)):
-        psus_qdev.append(Keysight_N8700_qdev(
+        psus_qdev.append(
+            Keysight_N8700_qdev(
                 dev=psus[i],
-                DAQ_trigger=DAQ_trigger.SINGLE_SHOT_WAKE_UP,
-                debug=DEBUG))
+                DAQ_trigger=DAQ_TRIGGER.SINGLE_SHOT_WAKE_UP,
+                debug=DEBUG,
+            )
+        )
 
     # DEBUG information
-    psus_qdev[0].worker_DAQ.debug_color  = ANSI.YELLOW
+    psus_qdev[0].worker_DAQ.debug_color = ANSI.YELLOW
     psus_qdev[0].worker_jobs.debug_color = ANSI.CYAN
-    psus_qdev[1].worker_DAQ.debug_color  = ANSI.GREEN
+    psus_qdev[1].worker_DAQ.debug_color = ANSI.GREEN
     psus_qdev[1].worker_jobs.debug_color = ANSI.RED
 
     for psu_qdev in psus_qdev: